ENH Deprecate "archive" for "hidden"
In all of the `civis.io` functions, change the `archive` parameter (this ran a callback which archived the script after completion) to `hidden`. Instead of archiving scripts when they finish, hide them on creation. This is what the `archive` parameter had been trying to accomplish. Setting `hidden` keeps the scripts from cluttering the UI while leaving them accessible to e.g. status queries from the API. We also save the latency of an extra API call on script completion.

The `archive` option will remain (with a deprecation warning) until v2 of the client.

This commit also adds the `hidden` option to the `query_civis` function, which didn't have it before.