PERF Cache swagger generation (#19)
The API endpoints returned by the Platform were already being cached, but parsing the results of this call takes > 0.5 seconds. Caching the result of the entire swagger-parsing process is memory-cheap and gives a 1000x decrease in time required to create identical client objects. Creation is now ~150 us, fast enough that we can re-create client objects as needed with no performance overhead.