BUG Ensure that polling thread shuts down when no longer needed
When a Python session shuts down, all threads are joined. That means that we need to wait for the polling thread to complete before the interpreter can shut down. It hadn't been noticable when polling was the primary way of getting results, but with the use of the notifications endpoint, it matters. I observed a very short job which used notifications wait 9.5 minutes to complete, because it had to wait for the `sleep` inside our polling thread to finish.

There's no way to shut down a thread in a ``ThreadPoolExecutor`` early, so replace it with our own thread. The ``cancel`` method lets us stop it early, and overloading ``join`` ensures that we never need to wait for a polling cycle to complete when exiting the interpreter. This also solves a problem I'd observed in testing, that the test session wouldn't exit until the poller shut down.