TST Conditionally select polling interval in tests
When we record tests with ``vcr``, we need to have normal polling intervals, so as not to break the Platform API. But when running tests later, we don't want to wait through the entire sequence of polling at the normal intervals. We had been solving this by patching some internals of `PollableResult`. That solution breaks when we swap in the new threading poller. Replace with conditional poll intervals. We need to remember to set `polling_interval` for all new tests, but in return we don't need to patch private methods of `PollableResult` (and it works with the new threading).