[CIVP-10592] Deprecate the api_key in favor of APIClient inputs (#46)
High-level functions and classes in this library all take `api_key` inputs and instantiate new `APIClient` objects, rather than allowing users to pass in an existing `APIClient` object. There are cases when it can be very helpful to be able to pass in existing client objects. The top two:
- Sometimes we have an `APIClient` but not an `api_key`. I've written `api_key=self.client._session.auth[0]` a couple of times in my code now, and that's not good practice.
- For testing, it's very helpful to be able to create one `APIClient` mock and pass that around. If all functions instantiate new objects, the `APIClient` needs to be patched separately in each module.

Add a deprecation decorator to make it easy to deprecate parameters and to find where we've deprecated things when it's time to remove them.