PERF Retry all HTTP verbs when we have a Retry-After header and an approved error
If we have a rate limit error (429) and a Retry-After header, it should always be safe to retry the request, regardless of the verb. The `Retry.is_retry` method starts by checking if the verb is on the whitelist (defaults to ['HEAD', 'GET', 'PUT', 'DELETE', 'OPTIONS', 'TRACE']) and immediately returns False if it is not. This PR changes our code so that we will retry on the `RETRY_AFTER_STATUS_CODES` ([413, 429, 503]) regardless of verb.

`urllib3` respects the "Retry-After" header starting at v1.19. We get `urllib3` from `requests`; `requests` updated `urllib3` to a compatible version at v2.12.0. Therefore this PR also needs to increase the version of `requests` which we require.