ENH Add ModelFuture
Add the new `civis.ml` namespace, and populate it with `ModelFuture`, a subclass of `CivisFuture`. The `ModelFuture` object knows how to find and download results from a CivisML model. It also transports errors from Civis and re-raises them in the local environment to make debugging easier.

This commit copies already-reviewed code from the development repository.