ENH ModelFuture recognize and raies MemoryErrors
Modify the `ModelFuture` to notice when the jobs it's connected to failed because they ran out of memory. If that happened, re-raise an appropriate error so that the user can handle it. We can't do anything about this on the Platform side -- Python dies abruptly without being able to write any outputs.