[CIVP-11427] BUG Improve thread safety in PollableResult (#92)
We have at least once instance of a bug which looks to be caused by non-threadsafe code. The most likely culprit is in the `PollableResult` stack. Add some extra locks to avoid possibly double-destroying anything. `cleanup` looks like it was always called from inside a function holding the lock, so it's probably fine. It doesn't hurt to add an extra lock. `_reset_polling_thread` wasn't behind a lock, and it was being used as a callback from `pubnub` -- perhaps this could have been a problem?