ENH ContainerFuture class to track container scripts
Add a `ContainerFuture` object which knows how to query container scripts, how to cancel container scripts, and how to retry container scripts. The `civis.ml.ModelFuture` can be a subclass of the `ContainerFuture`; models are a specialized container script.

This functionality (especially retries) adds more tooling for running concurrent jobs on the Civis Platform, and is a setup for executor classes coming in the next PR.