BUG Include joblib.load in wrapper function error handling
In the joblib backend remote runner script, the `joblib.load` call will fail if we try to load an object which refers to code that's not in our remote environment. That's an easy mistake to make, and we should make sure that users see an informative error from it. By moving the `load` call inside the try/finally block, we can record any exceptions from the `load` call and pass them back to the parent process.