ENH Serialize with cloudpickle
Switch from serializing with `joblib` to using `cloudpickle`. `joblib` is more efficient than pickle protocol 2 for storing numpy arrays, but pickle protocols 3 and 4 (available in Python 3) are just as good as `joblib` for array data. `cloudpickle` adds the ability to transport dynamically defined functions and classes, so that users don't need to create installable Python modules to run a single custom function in the Civis Platform.

Using protocol `pickle.HIGHEST_PROTOCOL` guarantees that we get the best protocol for our version of Python. This does assume that users have the same Python version on both ends of the `joblib.Parallel` call.