TST Run pandas and scikit-learn tests in Travis
`pandas` and `scikit-learn` are optional dependencies for the Python API client. We have tests which exercise these code paths, but the libraries aren't installed in the Travis testing environment, so we don't test the functions. This adds `pandas` and `sklearn` to Travis so that we're no longer skipping tests.

Upgrade the dev requirement of `vcrpy` to v1.11; the v1.10 release is not fully compatible with Python v3.6.