[CIVP-10885] ENH Retry connection errors in CLI (#117)
* ENH Retry connection errors in CLI

When retrieving the API spec in the Civis client CLI, use the same code as we use in the client library. This has the advantage of including automatic retries for connection failures, and automatic waiting for rate limit refreshes. Also modify `invoke` so that we include the same retries as we would if we were using the client library.

* REF DRY API connection

We were creating connections to the Civis API in multiple places. Centralize everything in a single `open_session` function, and use that everywhere.

* ENH Change CLI's User-Agent to "civis-cli"

Make the CLI's User-Agent string distinct from the Python client's User-Agent. The CLI will send a User-Agent string which starts with "civis-cli".