BUG Fully implement iterator protocol in PaginatedResponse
The `PaginatedResponse` provided iteration support, but didn't implement the `__next__` method. This meant that the object itself was not an iterator. A small tweak can turn it into an iterator itself. Since the `PaginatedResponse` is returned when users provide an `iterator=True` argument to list endpoints, it seems reasonable for them to expect the returned object is an iterator.