BUG Avoid IndexError on empty civis_to_csv queries (#135)
The `download_callback` assumed that the future it's attached to returned some output. If that weren't the case (either the query errored or succeeded with no results), then the user would see an `IndexError` display on the screen, but not halt program execution. This is confusing. Intercept the error and replace it with a warning.