TST Provide autospecced client mock for testing (#141)
* REF Move get_api_key to _utils

Promote the `get_api_key` function to the general `_utils` module so that it can be used outside of `civis`.

* REF Move default cache path to _resources

The CLI defined a default location for a cache of the API spec. Move that to the `_resources` module for general use.

* ENH Add cache_api_spec function

This function will make it easy for users to dump a cache of the API's JSON spec.

* ENH Add tests.mocks.create_client_mock

This function creates an autospecced mock suitable for use in local tests.

* REF Centralize TEST_SPEC definition

Point to one location for an API spec to be used in testing.

* TST Update cached API spec for tests

Only update the version with "channels"; most existing tests assume "channels" is not present.

* TST Default client mocks to use TEST_SPEC_ALL

Use the just-updated spec which ships with the package, rather than requiring users to download a spec themselves.

* MAINT Include API specs for testing with install

Now that we're supplying code to help users test their own code using the client, include the current API specs with new civis-python installs.