REF DRY polling thread restarts and always start after pubnub
Refactor the Futures code so that we handle thread starts and resets from the same function. Now `_begin_tracking` is the one function which sets (or resets) both status polling in the thread and PubNub channel listening. Make sure that we always subscribe to PubNub before we poll. This is important because otherwise we might poll, see a running job, and miss that the job completed in between polling and subscribing.

Potential problems fixed in this refactor:
- Now always polling after subscribing (the `CivisFuture` was polling before subscribing)
- When unpickling a `ModelFuture` object, set the polling function before creating the polling thread (we may have gotten intermittent errors when unpickling these objects before)