BUG Don't error when cancelling completed ContainerFutures (#165)
If you call `scripts.post_cancel` on a completed script, you get a `CivisAPIError` with a 404 status code. We'd gated this call behind an `if not self.done()`, but it's possible for jobs to complete before the done status updates. This is most prominent when using polling to update job status. This patch causes all 404 errors on `post_cancel` to return `False` for the cancelled status (since they obviously weren't cancelled). Non-404 `CivisAPIError` status codes will emit a warning and then return `False`.