REF DRY polling thread restarts and always start after pubnub (#167)
Refactor the Futures code so that we handle thread starts and resets from the same function. Now `_begin_tracking` is the one function which sets (or resets) both status polling in the thread and PubNub channel listening. Make sure that we always subscribe to PubNub before we poll. This is important because otherwise we might poll, see a running job, and miss that the job completed in between polling and subscribing.