PERF Don't keep an open session when we don't need to
We'd been holding open sockets in two places. This uses system file handles, which isn't a problem for a few things at a time, but which can eat up lots of resources if you're getting to hundreds of client objects and/or `CivisFuture`s.

First, we opened a `requests.Session` with every new client and kept it open. This wasn't getting closed even when the `APIClient` went out of scope. We can ensure that this session closes by only opening it when needed. There are performance implications here, but we send API requests infrequently enough that it's probably worth it to open a new session each time. I measured 0.1 ms to open and close a session in rough testing.

We're now storing the user's API key directly on the `APIClient` instance, so explicitly raise an error when pickling. This will ensure we don't accidentally leave keys where they shouldn't be.

The second place which keeps open sockets is PubNub, used in the `CivisFuture`. PubNub opens two sockets (I still don't know why there's two) when we subscribe, and it never gives them up, even after unsubscribing. PubNub does not provide a way to close the session in public methods, but it's possible to close the session directly on the private `_request_handler` attribute. I found that PubNub will then reopen one socket (not sure why), but I can avoid this by immediately removing the `Session` object. I then remove the `PubNub` object since it's no longer usable. (We would recreate it if we wanted to resubscribe anyway.) Note that `PubNub` is written with lots of circular references, so the Python garbage collector isn't going to be able to remove it. This represents a (small) memory leak.