BUG Start pubnub subscriptions in the past
If a SQL query fails, it can fail fast, sometimes faster than we can connect to our notifications channel. When we connect, request messages for the previous 30 seconds so that we can see if our job failed before we were able to finish setting up the connection to the notifications channel.