ENH Use feather to send data to CivisML, if possible
When users provide a `DataFrame` for model training or prediction, try first to send data with feather. This requires an additional package installation, but it's faster and it guarantees that we get the type correct for columns of categorical data. Require that we use feather for data with `pd.Categorical` columns, since we can't guarantee that we get the types correct when going through CSVs.