ENH Use feather to send data to CivisML, if possible (#200)
When users provide a `DataFrame` for model training or prediction, try first to send data with feather. This requires an additional package installation, but it's faster and it guarantees that we get the type correct for columns of categorical data. Require that we use feather for data with `pd.Categorical` columns, since we can't guarantee that we get the types correct when going through CSVs.

Existing CivisML releases don't support feather input. Make sure to send data as a CSV if the user is calling one of those versions.