ENH Public schema/tablename-splitting function
The `_robust_schema_table_split` is pretty useful, and users may wish to use it themselves. Promote to a public function with a clearer name. Also modify the function so that a single table name, without a schema, is a valid input. In such a case, return `None` for the schema name.