ENH Include parent job and run ID in children
It can be useful for child jobs to know which job created them. This injects environment variables into child jobs which give the job and run ID of the creator. Note that the joblib backend uses the `_ContainerShellExecutor` to launch its jobs, so this change affects it as well. Custom scripts will only have parent job and run IDs if the template provides that parameter. There's no way to inject environment variables which the template creator didn't explicitly allow for.