BUG Include JSONs in package data
We want to include JSON files with installations of the API client; there are JSON blobs with caches of API endpoints which are useful for testing. The `data_files` argument installs the JSON files when running `python setup.py install`, but they don't come when you run `pip install`. Setting `include_package_data=True` will include everything in the `MANIFEST.in`, including the JSON files.