Fix handling of empty responses from run cancellation endpoints.
The Scripts and Jobs run cancellation endpoints (e.g., DELETE /jobs/{job_id}/runs/{run_id})
return an empty response with a 202 error code.  The client currently raises
a CivisClientError("Unable to parse JSON from response", ...) in that case.
This fixes that so that users will be able to use the client with those endpoints.