BUG Fix test for new joblib
In v0.12, a joblib change meant that it relies on callback from the result object to decide when to terminate batched consumpution of inputs. Two of our tests mock the result object, which means that joblib's supplied callback doesn't get called. Set `n_jobs=1` so that `joblib` consumes the inputs all at once and doesn't need the callbacks.