Compatibility with joblib > 0.11 (#299)
* MAINT Don't call fixture directly

Use the fixture as a function input. Pytest complains if you call it directly.

* BUG Fix test for new joblib

In v0.12, a joblib change meant that it relies on callback from the result object to decide when to terminate batched consumpution of inputs. Two of our tests mock the result object, which means that joblib's supplied callback doesn't get called. Set `n_jobs=1` so that `joblib` consumes the inputs all at once and doesn't need the callbacks.

* ENH Update joblib

Relax requirement on the `joblib` version number, and modify our custom backend to include some new backend features.

* Suppress scikit-learn deprecation warning

Otherwise this warning shows up constantly, and confusingly for users.