BUG Use APIClient method name parsing for CLI (#316)
* TST Test for CLI warnings

We're getting a warning from CLI generation which indicates a problem. It should stay a warning in the code, so that changes to the Civis API don't expose bugs which render the CLI unusable, but this should cause our tests to fail until we fix it.

* BUG Use APIClient method name parsing for CLI

The CLI's name parsing was failing on /aliases/{object_type}/{alias}. Switch to using the version of this function which already existed in `civis.resources`, with some small munging to account for the CLI's naming style.