Add loosen_types to civis_file_to_tables call
A new parameter was added to the imports/files/csv endpoint to add the ability to "loosen" precisioned types like VARCHAR. We leverage this parameter when importing multiple files in a single import in order to accomodate differences in these lengths in the data. We also update the tests and the mock API spec to include the new parameter.