BUG: Add strip of semicolon to _get_headers
- As it stands now, if an error is tripped in _get_headers,
it silently fails. However, a simple fail can be invoked if the
sql being checked is terminated with a semicolon. So a strip of
the semicolon is added to cover a few more cases & get the
optimization of that routine. Because the entire failure is labeled
NOQA, tests aren't really possible. However, it was run with some
queries to verify it doesn't blow up.