BUG Correct sql_type doc casing and check input for typos
The client is looking for the `table_columns` parameter to be a list of dictionaries, each of which has a "name" key and which optionally also have a "sql_type" key. If Civis Platform needs to know types, and there is no "sql_type" key in any of the dictionaries, the client will silently replace a user-provided `table_columns` with an inferred `table_columns`. The docstring claims that the expected formatting is "sqlType", but "sqlType" keys will be ignored.

Fix the docstrings to use snake case instead of camel case, and add a check on the `table_columns` input to prevent mistakes or typos.