[CIVP-24324] FIX remove `_get_headers` in `read_civis_sql` (#415)
* BUG: Add strip of semicolon to _get_headers

- As it stands now, if an error is tripped in _get_headers,
it silently fails. However, a simple fail can be invoked if the
sql being checked is terminated with a semicolon. So a strip of
the semicolon is added to cover a few more cases & get the
optimization of that routine. Because the entire failure is labeled
NOQA, tests aren't really possible. However, it was run with some
queries to verify it doesn't blow up.

* REF remove _get_headers

* TST fix tests that don't use cassettes

* TST refactor vcr tests to mock-based ones

* MAINT update changelog

Co-authored-by: jacksonlee-civis <jacksonlee@civisanalytics.com>